added user authentication for all todos endpoints

+++ todo-app/routers/todos.py
@@ -4,6 +4,7 @@ from database import SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session
 from pydantic import BaseModel, Field
+from .auth import get_current_user
 
 
 
@@ -19,6 +20,8 @@ def get_db():
 
 db_dependency = Annotated[Session, Depends(get_db)]
 
+user_dependency = Annotated[dict, Depends(get_current_user)]
+
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
     description: str = Field(min_length=3, max_length=100)
@@ -26,13 +29,17 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get("/")
-def get_all_todos(db: db_dependency):
-    return db.query(Todos).all()
+def get_all_todos(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-def get_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+def get_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is not None:
         return todo_model
     else:
@@ -40,16 +47,22 @@ def get_todo(db: db_dependency, todo_id: int = Path(gt=0)):
     
 
 @router.post("/todo/add", status_code=status.HTTP_201_CREATED)
-def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get("id"))
 
     db.add(todo_model)
     db.commit()
 
      
 @router.put("/todo/edit/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-def edit_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int= Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+def edit_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int= Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found!")
     else:
@@ -63,10 +76,13 @@ def edit_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int= Path(g
 
 
 @router.delete("/todo/delete/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found!")
     else:
-        db.query(Todos).filter(Todos.id == todo_id).delete()
+        db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
         db.commit()
